refactor(widgets): вынесены типы в отдельный файл
вынесены типы в отдельный файл
